// packages/frontend/generated/api/schemas/notificationListResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * Chase Light API
 * GitHub リポジトリ監視サービス API - TypeScript + Hono + Zod + OpenAPI
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationListResponseData } from './notificationListResponseData';

export interface NotificationListResponse {
  success: boolean;
  data: NotificationListResponseData;
}